remove need for ArchiveView. Move ArchiveDestinationModifier into separate file
